refactor(AboutMe): move styles
